refactor : menuProduct 이동 (product -> menuProduct)
