+++ src/main/java/kitchenpos/order/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.order.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,6 +10,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
+import kitchenpos.product.domain.Product;
 
 import java.util.UUID;
