+++ src/test/java/kitchenpos/Fixtures.java
@@ -2,7 +2,7 @@ package kitchenpos;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.product.domain.MenuProduct;
+import kitchenpos.order.domain.MenuProduct;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
